claimable items migrated to the new design and template files deleted
