--- ADP.WebComponents/adp-web-components/src/components/vehicle-lookup/claim-temp-form.css
@@ -1,43 +1,0 @@
-:host {
-  all: initial !important;
-  display: block;
-}
-
-* {
-  font-family: Arial;
-}
-
-.lds-ripple {
-  @apply box-border absolute z-10 border-[4px] border-[#3071a9] rounded-full;
-  animation: lds-ripple 1s cubic-bezier(0, 0.2, 0.8, 1) infinite;
-}
-
-.lds-ripple:nth-child(2) {
-  animation-delay: -0.5s;
-}
-
-@keyframes lds-ripple {
-  0% {
-    width: 8px;
-    height: 8px;
-    opacity: 0;
-  }
-
-  4.9% {
-    width: 8px;
-    height: 8px;
-    opacity: 0;
-  }
-
-  5% {
-    width: 8px;
-    height: 8px;
-    opacity: 1;
-  }
-
-  100% {
-    width: 100%;
-    height: 100%;
-    opacity: 0;
-  }
-}
