+++ ADP.WebComponents/adp-web-components/src/components/vehicle-lookup/components/claimable-item-popover.tsx
@@ -1,8 +1,8 @@
-import cn from '~lib/cn';
 import { h, FunctionalComponent } from '@stencil/core';
 
 import { VehicleServiceItemDTO } from '~types/generated/vehicle-lookup/vehicle-service-item-dto';
 
+import cn from '~lib/cn';
 import ComponentLocale from '~lib/component-locale';
 
 import { TriangleIcon } from '~assets/triangle-icon';
